Conflicts:
	gns3server/modules/dynamips/nodes/router.py

+++ gns3server/modules/virtualbox/virtualbox_vm.py
@@ -29,11 +29,11 @@ import socket
 import asyncio
 
 from pkg_resources import parse_version
+from gns3server.utils.telnet_server import TelnetServer
 from .virtualbox_error import VirtualBoxError
 from ..nios.nio_udp import NIOUDP
 from ..nios.nio_nat import NIONAT
 from ..adapters.ethernet_adapter import EthernetAdapter
-from .telnet_server import TelnetServer  # TODO: port TelnetServer to asyncio
 from ..base_vm import BaseVM
 
 if sys.platform.startswith('win'):
@@ -65,6 +65,7 @@ class VirtualBoxVM(BaseVM):
         self._adapters = adapters
         self._ethernet_adapters = {}
         self._headless = False
+        self._acpi_shutdown = False
         self._enable_remote_console = False
         self._vmname = vmname
         self._use_any_adapter = False
@@ -73,17 +74,23 @@ class VirtualBoxVM(BaseVM):
 
     def __json__(self):
 
-        return {"name": self.name,
+        json = {"name": self.name,
                 "vm_id": self.id,
                 "console": self.console,
                 "project_id": self.project.id,
                 "vmname": self.vmname,
                 "headless": self.headless,
+                "acpi_shutdown": self.acpi_shutdown,
                 "enable_remote_console": self.enable_remote_console,
                 "adapters": self._adapters,
                 "adapter_type": self.adapter_type,
                 "ram": self.ram,
                 "use_any_adapter": self.use_any_adapter}
+        if self._linked_clone:
+            json["vm_directory"] = self.working_dir
+        else:
+            json["vm_directory"] = None
+        return json
 
     @asyncio.coroutine
     def _get_system_properties(self):
@@ -142,7 +149,7 @@ class VirtualBoxVM(BaseVM):
 
         yield from self._get_system_properties()
         if "API version" not in self._system_properties:
-            raise VirtualBoxError("Can't access to VirtualBox API Version")
+            raise VirtualBoxError("Can't access to VirtualBox API version:\n{}".format(self._system_properties))
         if parse_version(self._system_properties["API version"]) < parse_version("4_3"):
             raise VirtualBoxError("The VirtualBox API version is lower than 4.3")
         log.info("VirtualBox VM '{name}' [{id}] created".format(name=self.name, id=self.id))
@@ -161,6 +168,19 @@ class VirtualBoxVM(BaseVM):
         vm_info = yield from self._get_vm_info()
         if "memory" in vm_info:
             self._ram = int(vm_info["memory"])
+
+    @asyncio.coroutine
+    def check_hw_virtualization(self):
+        """
+        Returns either hardware virtualization is activated or not.
+
+        :returns: boolean
+        """
+
+        vm_info = yield from self._get_vm_info()
+        if "hwvirtex" in vm_info and vm_info["hwvirtex"] == "on":
+            return True
+        return False
 
     @asyncio.coroutine
     def start(self):
@@ -196,20 +216,29 @@ class VirtualBoxVM(BaseVM):
         if self._enable_remote_console and self._console is not None:
             self._start_remote_console()
 
+        if (yield from self.check_hw_virtualization()):
+            self._hw_virtualization = True
+
     @asyncio.coroutine
     def stop(self):
         """
         Stops this VirtualBox VM.
         """
 
+        self._hw_virtualization = False
         self._stop_remote_console()
         vm_state = yield from self._get_vm_state()
         if vm_state == "running" or vm_state == "paused" or vm_state == "stuck":
-            # power off the VM
-            result = yield from self._control_vm("poweroff")
-            log.info("VirtualBox VM '{name}' [{id}] stopped".format(name=self.name, id=self.id))
-            log.debug("Stop result: {}".format(result))
+            if self.acpi_shutdown:
+                # use ACPI to shutdown the VM
+                result = yield from self._control_vm("acpipowerbutton")
+                log.debug("ACPI shutdown result: {}".format(result))
+            else:
+                # power off the VM
+                result = yield from self._control_vm("poweroff")
+                log.debug("Stop result: {}".format(result))
 
+            log.info("VirtualBox VM '{name}' [{id}] stopped".format(name=self.name, id=self.id))
             # yield from asyncio.sleep(0.5)  # give some time for VirtualBox to unlock the VM
             try:
                 # deactivate the first serial port
@@ -318,6 +347,7 @@ class VirtualBoxVM(BaseVM):
                     if nio and isinstance(nio, NIOUDP):
                         self.manager.port_manager.release_udp_port(nio.lport, self._project)
 
+        self.acpi_shutdown = False
         yield from self.stop()
 
         if self._linked_clone:
@@ -388,6 +418,30 @@ class VirtualBoxVM(BaseVM):
         else:
             log.info("VirtualBox VM '{name}' [{id}] has disabled the headless mode".format(name=self.name, id=self.id))
         self._headless = headless
+
+    @property
+    def acpi_shutdown(self):
+        """
+        Returns either the VM will use ACPI shutdown
+
+        :returns: boolean
+        """
+
+        return self._acpi_shutdown
+
+    @acpi_shutdown.setter
+    def acpi_shutdown(self, acpi_shutdown):
+        """
+        Sets either the VM will use ACPI shutdown
+
+        :param acpi_shutdown: boolean
+        """
+
+        if acpi_shutdown:
+            log.info("VirtualBox VM '{name}' [{id}] has enabled the ACPI shutdown mode".format(name=self.name, id=self.id))
+        else:
+            log.info("VirtualBox VM '{name}' [{id}] has disabled the ACPI shutdown mode".format(name=self.name, id=self.id))
+        self._acpi_shutdown = acpi_shutdown
 
     @property
     def enable_remote_console(self):
@@ -504,7 +558,7 @@ class VirtualBoxVM(BaseVM):
         """
         Returns either GNS3 can use any VirtualBox adapter on this instance.
 
-        :returns: index
+        :returns: boolean
         """
 
         return self._use_any_adapter
@@ -520,7 +574,7 @@ class VirtualBoxVM(BaseVM):
         if use_any_adapter:
             log.info("VirtualBox VM '{name}' [{id}] is allowed to use any adapter".format(name=self.name, id=self.id))
         else:
-            log.info("VirtualBox VM '{name}' [{id}] is not allowd to use any adapter".format(name=self.name, id=self.id))
+            log.info("VirtualBox VM '{name}' [{id}] is not allowed to use any adapter".format(name=self.name, id=self.id))
         self._use_any_adapter = use_any_adapter
 
     @property
@@ -574,10 +628,11 @@ class VirtualBoxVM(BaseVM):
 
         # check the maximum number of adapters supported by the VM
         vm_info = yield from self._get_vm_info()
-        chipset = vm_info["chipset"]
         maximum_adapters = 8
-        if chipset == "ich9":
-            maximum_adapters = int(self._system_properties["Maximum ICH9 Network Adapter count"])
+        if "chipset" in vm_info:
+            chipset = vm_info["chipset"]
+            if chipset == "ich9":
+                maximum_adapters = int(self._system_properties["Maximum ICH9 Network Adapter count"])
         return maximum_adapters
 
     def _get_pipe_name(self):
